+++ src/app/(components)/BackToTopButton.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { ArrowUpIcon } from '@heroicons/react/24/outline'
+import { ArrowUpIcon } from '@heroicons/react/24/solid'
 
 export default function BackToTopButton() {
   function handleBackToTopClick() {
@@ -10,7 +10,7 @@ export default function BackToTopButton() {
   return (
     <div className="flex items-center justify-start">
       <button
-        className="flex w-auto items-center justify-center gap-1 rounded-md border-2 border-transparent  bg-yellow-500 p-2 text-sm font-bold text-black drop-shadow-md hover:border-yellow-300"
+        className="flex w-auto items-center justify-center gap-1 rounded-md border-2 border-black bg-yellow-500 p-2 text-sm font-bold text-black drop-shadow-lg hover:border-yellow-300"
         onClick={handleBackToTopClick}
       >
         <ArrowUpIcon className="h-5 w-5" />