+++ src/app/(components)/Dialog.tsx
@@ -3,7 +3,7 @@
 import { Fragment } from 'react'
 import { Dialog as BaseDialog, Transition } from '@headlessui/react'
 import { cn } from '@/app/(lib)/utils'
-import { XMarkIcon } from '@heroicons/react/24/outline'
+import { XMarkIcon } from '@heroicons/react/24/solid'
 
 interface DialogProps {
   children?: React.ReactNode