+++ src/app/(components)/Logo.tsx
@@ -6,7 +6,7 @@ export default function Logo({ showUrl = false }: { showUrl?: boolean }) {
     <Link href="/" className="-m-1.5 flex items-center justify-start p-1.5">
       <Image
         className="mr-2"
-        src={`https://${process.env.NEXT_PUBLIC_IMAGE_URL}/logo-sm.png`}
+        src="https://remnant2toolkit.b-cdn.net/logo-sm.png?width=32&height=32"
         alt="Remnant 2 Toolkit logo, a purple and yellow toolbox."
         height={32}
         width={32}