+++ src/app/builder/(components)/BuilderButton.tsx
@@ -85,7 +85,7 @@ export default function BuilderButton({
       >
         {item && (
           <Image
-            src={`https://${process.env.NEXT_PUBLIC_IMAGE_URL}${item.imagePath}`}
+            src={`https://${process.env.NEXT_PUBLIC_IMAGE_URL}${item.imagePath}?width=${imageSize.width}&height=${imageSize.height}`}
             alt={`${item.name} icon`}
             width={imageSize.width}
             height={imageSize.height}