+++ src/app/builder/(components)/BuilderName.tsx
@@ -1,4 +1,4 @@
-import { PencilIcon } from '@heroicons/react/24/outline'
+import { PencilIcon } from '@heroicons/react/24/solid'
 import { useState } from 'react'
 
 export default function BuilderName({