+++ src/app/builder/[buildId]/page.tsx
@@ -10,7 +10,7 @@ import useBuildActions from '../../(hooks)/useBuildActions'
 import { ActionButton } from '../(components)/ActionButton'
 import ToCsvButton from '@/app/(components)/ToCsvButton'
 import { useIsClient } from 'usehooks-ts'
-import { useRef } from 'react'
+import { useEffect, useRef } from 'react'
 import { useSession } from 'next-auth/react'
 import PageHeader from '@/app/(components)/PageHeader'
 import { ExtendedBuild, isErrorResponse } from '@/app/(types)'
@@ -20,6 +20,7 @@ import ImageDownloadLink from '../(components)/ImageDownloadLink'
 import { addVoteForBuild, removeVoteForBuild } from '../actions'
 import { toast } from 'react-toastify'
 import { useRouter } from 'next/navigation'
+import { useLocalStorage } from '@/app/(hooks)/useLocalStorage'
 
 export default function Page({
   params: { extendedBuild },
@@ -40,10 +41,25 @@ export default function Page({
     handleDuplicateBuild,
     handleEditBuild,
     handleImageExport,
+    handleScrollToDetailedView,
     handleReportBuild,
   } = useBuildActions()
 
+  const { builderStorage, setBuilderStorage } = useLocalStorage()
+
+  useEffect(() => {
+    if (!builderStorage.tempBuildId) return
+    setBuilderStorage({
+      ...builderStorage,
+      tempBuildId: null,
+      tempDescription: null,
+      tempIsPublic: null,
+      tempCreatedById: null,
+    })
+  })
+
   const buildContainerRef = useRef<HTMLDivElement>(null)
+  const detailedViewContainerRef = useRef<HTMLDivElement>(null)
 
   // Need to convert the build data to a format that the BuildPage component can use
   const buildState = extendedBuildToBuildState(extendedBuild)
@@ -70,24 +86,28 @@ export default function Page({
             id="actions-column"
             className="flex min-w-full flex-col justify-between sm:min-w-[100px]"
           >
-            <div id="actions" className="flex flex-col gap-2">
+            <div
+              id="actions"
+              className="grid grid-cols-2 gap-2 sm:flex sm:flex-col sm:gap-2"
+            >
+              <div className="col-span-full">
+                <ActionButton.ExportImage
+                  imageExportLoading={imageExportLoading}
+                  onClick={() =>
+                    handleImageExport(
+                      buildContainerRef.current,
+                      `${buildState.name}.png`,
+                    )
+                  }
+                />
+              </div>
+
               {session && session.user?.id === buildState.createdById && (
                 <ActionButton.EditBuild
                   onClick={() => handleEditBuild(buildState)}
                 />
               )}
-              <ActionButton.DuplicateBuild
-                onClick={() => handleDuplicateBuild(buildState)}
-              />
-              <ActionButton.ExportImage
-                imageExportLoading={imageExportLoading}
-                onClick={() =>
-                  handleImageExport(
-                    buildContainerRef.current,
-                    `${buildState.name}.png`,
-                  )
-                }
-              />
+
               <ActionButton.CopyBuildUrl
                 onClick={() =>
                   handleCopyBuildUrl(
@@ -96,13 +116,23 @@ export default function Page({
                   )
                 }
               />
+              <ActionButton.DuplicateBuild
+                onClick={() => handleDuplicateBuild(buildState)}
+              />
               <ToCsvButton
                 data={csvBuildData.filter((item) => item?.name !== '')}
                 filename={`remnant2_builder_${buildState.name}`}
+                label="Export to CSV"
+              />
+
+              <ActionButton.ShowDetailedView
+                onClick={() =>
+                  handleScrollToDetailedView(detailedViewContainerRef.current)
+                }
               />
               {session?.user && (
                 <>
-                  <hr className="my-4 border-gray-500" />
+                  <hr className="my-4 hidden border-gray-500 sm:visible" />
 
                   <div className="flex w-full flex-col items-center justify-center gap-4">
                     <TotalUpvotes
@@ -110,35 +140,33 @@ export default function Page({
                       isMember={buildState.isMember}
                     />
 
-                    <ActionButton.Vote
-                      active={buildState.upvoted}
-                      onClick={async () => {
-                        const newVote = !buildState.upvoted
+                    <div className="my-4 flex flex-row items-center justify-center gap-x-4 sm:my-0 sm:flex-col sm:items-start sm:gap-x-0 sm:gap-y-2">
+                      <ActionButton.Vote
+                        active={buildState.upvoted}
+                        onClick={async () => {
+                          const newVote = !buildState.upvoted
 
-                        const response = newVote
-                          ? await addVoteForBuild(
-                              JSON.stringify({ buildId: extendedBuild.id }),
+                          const response = newVote
+                            ? await addVoteForBuild(
+                                JSON.stringify({ buildId: extendedBuild.id }),
+                              )
+                            : await removeVoteForBuild(
+                                JSON.stringify({ buildId: extendedBuild.id }),
+                              )
+
+                          if (isErrorResponse(response)) {
+                            console.error(response.errors)
+                            toast.error(
+                              'Error voting for build. Please try again later.',
                             )
-                          : await removeVoteForBuild(
-                              JSON.stringify({ buildId: extendedBuild.id }),
-                            )
-
-                        if (isErrorResponse(response)) {
-                          console.error(response.errors)
-                          toast.error(
-                            'Error voting for build. Please try again later.',
-                          )
-                        } else {
-                          toast.success(response.message)
-                          buildState.upvoted = newVote
-                          buildState.totalUpvotes = response.totalUpvotes ?? 1
-                          router.refresh()
-                        }
-                      }}
-                    />
-                  </div>
-
-                  <hr className="my-4 border-gray-500" />
+                          } else {
+                            toast.success(response.message)
+                            buildState.upvoted = newVote
+                            buildState.totalUpvotes = response.totalUpvotes ?? 1
+                            router.refresh()
+                          }
+                        }}
+                      />
 
                   <div className="flex w-full flex-col items-center justify-center gap-4">
                     <ActionButton.ReportBuild
@@ -169,11 +197,11 @@ export default function Page({
             />
           </div>
         </div>
-        <div className="mt-12 flex w-full flex-col items-center justify-center gap-2">
-          <DetailedBuildView
-            buildState={buildState}
-            isScreenshotMode={isScreenshotMode}
-          />
+        <div
+          className="mt-12 flex w-full flex-col items-center justify-center gap-2"
+          ref={detailedViewContainerRef}
+        >
+          <DetailedBuildView buildState={buildState} />
         </div>
       </div>
     </>