+++ src/app/profile/favorited-builds/page.tsx
@@ -44,9 +44,6 @@ export default function Page() {
 
   return (
     <div className="mx-auto w-full bg-black py-10">
-      <PageActions>
-        <BackToTopButton />
-      </PageActions>
       <div className="px-4 sm:px-6 lg:px-8">
         <div className="w-full sm:flex sm:items-center">
           <div className="sm:flex-auto">